feat
- 增加方块硬度API
- 删除Register注解，改为Enable内注册
- 优化了无用的import

+++ api/src/main/java/io/github/moyusowo/neoartisanapi/api/block/state/ArtisanCropState.java
@@ -9,6 +9,8 @@ import io.github.moyusowo.neoartisanapi.api.util.BuilderFactoryUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.Optional;
+
 /**
  * 作物型方块状态定义。
  *
@@ -41,7 +43,7 @@ public interface ArtisanCropState extends ArtisanBaseBlockState {
 
     @Override
     @Nullable
-    default Integer getHardness() {
+    default Float getHardness() {
         return null;
     }
